refactor(commands): extract command list, basic commands and pseudo commands in new files
