+++ index.js
@@ -14,306 +14,14 @@ import {
 import { randInt } from "./src/utils/randInt.js";
 import { smileys } from "./src/constantes/smileys.js";
 import { chanIds } from "./src/constantes/chanIds.js";
+import { commandes, cpseudo, cinvis } from "./src/commandes/commandList.js";
+import { basicCommands } from "./src/commandes/basicCommands.js";
+import { pseudoCommands } from "./src/commandes/pseudoCommands.js";
 
 // Variables globales
 const hll = new Date("2022", "07", "09", "20", "08", "00");
 let generalChan; // Objet contenant les réfs vers le canal #general
-const chanBeauf = `<#${chanIds.beauf}>`;
 
-// Commandes
-const commandes = [
-  "!help",
-  "!forum",
-  "!trombi",
-  "!site",
-  "!requete",
-  "!d + nombre",
-  "!everyone",
-  "!echo",
-  "!cui",
-  "!nuclear",
-  "!cyanure",
-  "!bleus",
-  "!actualité",
-  "!log",
-  "!classement",
-  "!commandes",
-  "!rich",
-  "!profil _nomDiscord",
-  "!penis",
-  "!oreoplz",
-  "!parapluie",
-  "!film",
-  "!license",
-];
-const cpseudo = [
-  "!Bernard",
-  "!Golgoth",
-  "!Maelg",
-  "!MacKay",
-  "!Roventa",
-  "!Sibaal",
-  "!Cleme",
-  "!Khran",
-  "!Drac",
-  "!Dragon",
-  "!Matsam",
-  "!Shaggyz",
-  "!Uff",
-  "!Hartyom",
-  "!Tephal",
-  "!Cerfpihier",
-  "!Lilith",
-  "!Loko",
-  "!WassaW",
-  "!Nathan",
-  "!Zomzom",
-  "!Pellaeon",
-  "!Roystar",
-  "!Dikryl",
-  "!Alexis",
-  "!Ganapati",
-  "!Mad",
-  "!Mzboub",
-  "!Freez",
-];
-const cinvis = ["!membres", "!logs"];
-// commandes => Pour les commandes basiques
-// cpseudo => Pour les commandes liées au pseudo
-// cinvis => Pour les commandes qui n'apparaissent pas dans le !help
-const asw = {
-  help: [
-    `**Commandes disponibles :**\n\`\`\`${commandes.join(
-      "\n"
-    )}\`\`\`\n_!membres_ pour plus de commandes\n_!help commande_ pour plus d'information sur une commande`,
-  ],
-  membres: [
-    `**Commandes disponibles :**\n\`\`\`${cpseudo.join(
-      "\n"
-    )}\`\`\`\n_!help_ pour plus de commandes`,
-  ],
-  forum: ["http://kirov-meta.forumactif.org/"],
-  trombi: [
-    "Pour accéder au trombi, demandez l'accès à Cleme ou Khran et postez une photo !",
-  ],
-  site: ["https://bernard-site.herokuapp.com/"],
-  everyone: ["Normalement ce message n'est pas censé apparaître :calim:"],
-  echo: [
-    "Tu t'es cru dans une caverne ?!",
-    "Syncope générale provoquée par tts imminente.",
-  ],
-  cui: ["Cui !", "Tu me prends pour un vulgaire moineau?"],
-  nuclear: ["Alex The Autist vous détruira tous !"],
-  cyanure: [`J'attends toujours que Matsam me crée ${smileys.sob}`],
-  bleus: [
-    "On est les champions ! On est les champions ! On est... on est... on est les champions ! ♫",
-    "Allez les bleus !",
-    "1998-2018 !",
-    "Et 1... Et 2... Et 3... et 4-2 !",
-  ],
-  actualité: [
-    "Google est ton ami !",
-    "Hartyom n'a toujours pas fini de se plaindre",
-    "On est en guerre.",
-  ],
-  log: [
-    "***Last update : 28/11/2020 19h00***\n\nBernard souhaite maintenant l'anniversaire des actifs du discord !\nAjout des commandes !film, !Mzboub, !Mad, !Freez.\nMise à jour des commandes !Alexis, !Maelg, !WassaW, !Zomzom, !Roystar, !Uff, !Pellaeon, !Cleme, !Hartyom et !actualité.\n\n_!logs_ pour plus de logs",
-  ],
-  logs: ["https://bernard-site.herokuapp.com/logs"],
-  penis: ["8=D"],
-  oreoplz: [smileys.oreo],
-  parapluie: [
-    `||${smileys.koukou}||||${smileys.mhh}||||${smileys.drop}||||${smileys.umbrella}||`,
-  ],
-  film: [
-    "You shit on my garden ?!",
-    `Ils sont entrain de manger ma maman${smileys.sadcat}`,
-    "Don't piss on the hospitality",
-    "Vous êtes un koala tout à fait décent",
-    "La rivière de nos larmes coule après le koala",
-    "Beaux yeux belles couilles",
-  ],
-  license: [
-    "https://github.com/MaeIg/Libert-Accabl-e-Bot/blob/master/LICENSE.md",
-  ],
-  Bernard: ["C'est moi !"],
-  Golgoth: ["Tu parle à ki ?", "tu me parle à moi?", "j'suis un gangster"],
-  Maelg: [
-    "Miou !",
-    "J'aime pas Despacito, je l'ai juste jouée 50 fois sur osu!",
-    ":3",
-    "Tu veux voir mon babobab ?",
-    "Ma copine c'est encore pire elle peut les écarter à fond",
-    "En vrai jsuis sûr que j'aimerais bien la soupe de bienvenue",
-    "Pourquoi tout ce qui tourne autour de Supken est d'une débilité incommensurable ?",
-    "Wesh c'est mon île pas un logement social",
-  ],
-  MacKay: [
-    "L'humanité est pourrie l'amour est une illusion",
-    "Je vous suis infiniment supérieur ",
-    "Loué soit le grand MacKay",
-    "Ma vie pour MacKay",
-    "Et si on abandonnait Matsang pour la quinzième fois ?",
-    "Les éclaireurs : poules mouillées ou tafioles?",
-  ],
-  Roventa: [
-    "Je t'aime... mais lâche-moi bordel !",
-    "Cool une vodka, je ne serais pas obligé de boire mon eau pour finir l'expé !",
-    "Les russes ont triché l'Égypte devait gagner",
-    "t'habite ou ?",
-    "T'as du style, épouse-moi !",
-    "Cool un RP ! Jump antipictoquoi ?",
-    "Ça m énerve pcq je ne suis pas à la hauteur de tout ce que tu nous apporte",
-    "Une fille qui joue a hordes, ca ne peut pas etre qq'un de manipulateur et d' égocentrique",
-    "Allez allez vous faites moins les malines quand vous nous mangez le sexe",
-  ],
-  Sibaal: ["Et en plus tu m'traîte de Golgoth."],
-  Cleme: [
-    "Et c'est sur ces mots que toute envie révolutionnaire s'éteint chez le peuple.",
-    "Le mandat de la présidente Cleme était incontestable et sa puissance sans limite.",
-    "Île-de-France? Tu veux dire Paris",
-    "Y a que Paint de vrai",
-    chanBeauf,
-    `Je préfère le gratin dauphinois à la bite ${smileys.okhand}`,
-    "CLIQUE SUR LES CHARIOTS SALE CON",
-    "Tant mieux qu'elle soit morte ils devraient tous mourir ces sales americains du Sud",
-  ],
-  Khran: [
-    "Un vrai ETListe abandonne son stuff dans le désert !",
-    "Meurt dans le désert",
-  ],
-  Drac: ["+37 !", "J'ai disparu en 1937 !"],
-  Dragon: [
-    "Coup critique !",
-    "Bon je vais me coucher, bonne journée !",
-    "Oh, j'ai encore fait un 20 aux dés !",
-  ],
-  Matsam: [
-    `Hartychou ${smileys.heart}`,
-    `Paint ${smileys.heart}`,
-    `Ragots ${smileys.heart}`,
-    `Smash Bros ${smileys.heart}`,
-    "Maelg le dieu de l'anti-picto",
-    `Jpp de ma vie ${smileys.calim}`,
-    `Je vais en faire un meme${smileys.mat100}`,
-    "Pourquoi on est J2 ? On jump pas ce soir ?",
-    smileys.oppression,
-    'Perso j\'ai deux manières de plier, une manière "propre" et une manière "mi-propre". Comme ça, quand je prends le vêtement, en fonction de la manière dont il est plié, je sais comment il est',
-    "LES GENS QUI CRIE TOUS DES CON",
-    "ALERTE CANDICE A APPORTÉ...",
-    "JE SUIS SUR MON ÎLE T'ES DANS TA BILE",
-    "J'entends ça je tombe par terre sans vouloir être dramatique",
-    "Vous avez déjà éternué en rigolant ?",
-    "Avant j'étais un gros fayot et maintenant je suis juste une crotte inactive, j'aime beaucoup",
-  ],
-  Shaggyz: [
-    "Quel fléau la canicule parisienne, heureusement que je suis dans ma piscine au soleil",
-    "Je suis mort dehors J1 mais c'est pas ma faute je dormais chez ma copine",
-    "Je suis mort dehors mais c'est pas ma faute c'était la finale",
-  ],
-  Uff: ["Force et honneur !", "Génial les smoothies au foutre !"],
-  Hartyom: [
-    `JE SUIS PRISONNIER DU CANAL AUDIO, O SECOUR !${smileys.calim}${smileys.calim}${smileys.calim}`,
-    "VOUS CONNAISSEZ LE POINT COMMUN ENTRE UFF ET LE FROMAGE DE MACKAY ? ON EN ENTEND BEAUCOUP PARLER MAIS ON LES VOIT JAMAIS HAHAHAHAHAHAHAHHAAH",
-    "Y'a moyen de pécho dans la méta alors ? Ptain c'est de mieux en mieux",
-    "Il ne peut en rester AUCUN. Je ne laisserai aucune de ces racailles faire la loi dans cette ville. NOT UNDER MY WATCH",
-    "Ger une revandication.",
-    "JE VOUS BEZ",
-    "GREVEGENERALE",
-    "Quitte à être zoophile tu pourrais te renseigner",
-  ],
-  Tephal: [
-    "tg",
-    "Les meilleurs articles pour pécho s'achètent chez Moulinooooooooot.",
-    "VTFF",
-    "Ce sera répété, amplifié et déformé ou inversement réciproquement dans l'autre sens, au choix",
-  ],
-  Cerfpihier: ["Écoute petit..."],
-  Lilith: [
-    `Je suis aussi douée avec ma langue, on peut bien s'entendre${smileys.feu}`,
-    `Je suis très flexible sur les horaires (pas que sur les horaires ${smileys.smile} )`,
-    "J'étais bourrée xD",
-    "Là où je passe, les PMV trépassent",
-    `Hordes va disparaître en 2020 et j'aurai pas de titre sm ${smileys.sadcat}`,
-    "Elle est bonne la coquine ?",
-    "T'es plus mon crush",
-    "Jcrois qu'ils ont parlé de moi en rujono hier\nEn ruinoln\nEn Réunion.",
-  ],
-  Loko: [
-    "On est des gentils ou des mechants ?",
-    "Bravo Hartyom t'es le plus fort de la ville",
-    "Ça freine mes débuts de serial killer",
-    "allez violé moi",
-    "Je viens de recompter tu m'as deja sauvé 3 fois donc la prochaine c'est bon tu peux me laisser mourir",
-    "Manque plus qu'on me viole dans le désert et je suis au top",
-    `Shaggyz Cleme je vous love ${smileys.heart} ${smileys.heart} coeur sur vous`,
-  ],
-  WassaW: [
-    "J AI LU PENDANT 72 HEURES NON STOP AVEC UN SOIR DANS LA NUIT BLANCHE",
-    "Fais un pavé s'il te plait",
-    "NEGATIF IL EST TOUJOURS SOUS PLASTIQUE ON Y TOUCHE PAS",
-    "AHAHAHAHAH",
-    "Je peux manger du fromage blanc en attendant que tu comprennes ?",
-    "Il vaut mieux aller élever des escargots à Auxerre avec un doctorat plutôt que sans",
-    "Je n'arrive pas à me concentrer de tomate",
-    "Non je vote pas sur Allocine j'aime pas le jaune je deteste ce site",
-    `${smileys.heart} Vous êtes mes petits gėnies du DiDi ${smileys.heart}`,
-    "Ils s'auto dépriment sans moi !",
-    "Pour que deux Moi fassent un Nous il faut que les Moi soi soi",
-    'Toujours quand je suis sérieux on me "dé-sérieux"',
-    "On veut juste mon corps parce que je suis bonne, grâce à vous j'ai compris HAHAHAHAHAH",
-    "j ai fait des photos mais c'est pas très joli",
-  ],
-  Nathan: [
-    "@everyøne J'ai besoin de garants pour ma sortie de prison ! AIDEZ MOI",
-  ],
-  Zomzom: [
-    `Je tiens à dire que "${smileys.oppression}" c'est opressant mais que le "wesh" a la fin de ta phrase m'a mis au fond`,
-    "C'est à force de lire les magouilles de Shaggyz, j'apprends",
-    "Bah non a une soirée j'avais trop bu je l'ai embrassé, la boulette",
-    "Bonjour la méta, je suis à un hâle et malgré l'heure bon taudive, j'ai bu bien :D Désolé de ma mort de noob lors du précédent CM :'( ça m' artiste !",
-    "Vous me faites trop rire je vous aime aussi haha.",
-    "Je viens pas bourré ici okay",
-    "Le masque c'est comme l'apres shampooing : c'est sous la douche",
-    "L'absence d'info est une info tant que c'est communiqué.",
-    "Ce qui est plus chelou c'est ton aviq sur les viuex et le corona",
-    "INHUSTICR",
-  ],
-  Pellaeon: [
-    '"caquelon", c\'est un poulet travesti qui caquette ?',
-    'Sous word il y a deux "espaces", un avec "espace", un second généré automatique, comme quand tu met des "guillements" et qui se place tout seul',
-    "Je tente de mettre en demeure sfr de revoir toutes les resiliations de l'année devant la repression des fraudes ? X)",
-    "George Clone né ?",
-    "Arf, je décrète officiellement la creation du pays des imaginaire alors clochette !",
-    "Comme une expé : avec un fouet, de la volonté et des MSE !",
-    "La logique et le bon sens n'ont pas de place dans cette maison.",
-    "Plus de licornes et moins d'argent au quotidien ?",
-    "Vous êtes Rabin en informations mademoiselle. Mais ça correspond au personne, j'apprécie la logique !",
-    "C'est du pas toi local ?",
-    "Même s'il est ironique ee moins de vue de ne préparer l'armée qu'a un changement annuel, de moins point de vue xD",
-    "Ca dépend, si son signe astrologique est le cancer, c'est p'têt une version française Symfony",
-    "Que nenniema mie",
-  ],
-  Roystar: [
-    "Ptain j'ai mis mon caleçon à l'envers ce matin jme disais bien que c'était bizarre",
-    `Ah mais la prochaine fois je ramène du champomy moi vous m'avez fait trop peur jme souviens pas de tout ${smileys.scream}`,
-    "t ban",
-  ],
-  Dikryl: ["Salutérus, ça vagin ?"],
-  Alexis: [
-    `On a déjà vomi dans mon duvet et dans mes oreilles c'était vraiment pas super ${smileys.sadcat}`,
-    'Voulez vous écouter "on va niquer ce soir" ?',
-    "TU CROIS QUE TU COMPRENDS PAS MAIS EN FAIT TU COMPRENDS PAS ENCORE MOINS",
-  ],
-  Ganapati: [
-    "Merci Harty, n'oublions pas la \"recrue décence\", à l'heure où la recrudescence du molestage et du bizutage bat son plein ! xD",
-  ],
-  Mad: ["Personne ira visiter ton île.", smileys.humpf],
-  Mzboub: ["!menbre"],
-  Freez: [
-    "la sélection naturelle c'est le fait qu'un organisme survivre dans un milieu et pas un autre organisme, selon moi",
-  ],
-};
 const helpInfo = {
   help: "**Utilisation**\n```!help _Commande```\n**Description**\n```La commande help donne des informations sur l'utilisation de la commande spécifiée. Si aucune commande n'est spécifiée, elle donne la liste des commandes disponibles.```",
   membres:
@@ -508,12 +216,13 @@ bot.on("message", (msg) => {
   ) {
     msg.channel.send(`Bah c'est toi idiot ${smileys.calim}`);
     newCommand(msg.author, txt);
-  } else if (
-    commandes.indexOf(txt) !== -1 ||
-    cpseudo.indexOf(txt) !== -1 ||
-    cinvis.indexOf(txt) !== -1
-  ) {
-    const L = asw[txt.substr(1)];
+  } else if (commandes.indexOf(txt) !== -1 || cinvis.indexOf(txt) !== -1) {
+    const L = basicCommands[txt.substr(1)];
+    msg.channel.send(L[randInt(L.length)]);
+
+    newCommand(msg.author, txt);
+  } else if (cpseudo.indexOf(txt) !== -1) {
+    const L = pseudoCommands[txt.substr(1)];
     msg.channel.send(L[randInt(L.length)]);
 
     newCommand(msg.author, txt);
