+++ src/commandes/commandList.js
@@ -1,0 +1,63 @@
+// commandList => For basic commands (text/function)
+// pseudoCommandList => For commands linked to a user
+// invisibleCommandList => For commands that should not appear in !help
+
+const commandes = [
+  "!help",
+  "!forum",
+  "!trombi",
+  "!site",
+  "!requete",
+  "!d + nombre",
+  "!everyone",
+  "!echo",
+  "!cui",
+  "!nuclear",
+  "!cyanure",
+  "!bleus",
+  "!actualité",
+  "!log",
+  "!classement",
+  "!commandes",
+  "!rich",
+  "!profil _nomDiscord",
+  "!penis",
+  "!oreoplz",
+  "!parapluie",
+  "!film",
+  "!license",
+];
+const cpseudo = [
+  "!Bernard",
+  "!Golgoth",
+  "!Maelg",
+  "!MacKay",
+  "!Roventa",
+  "!Sibaal",
+  "!Cleme",
+  "!Khran",
+  "!Drac",
+  "!Dragon",
+  "!Matsam",
+  "!Shaggyz",
+  "!Uff",
+  "!Hartyom",
+  "!Tephal",
+  "!Cerfpihier",
+  "!Lilith",
+  "!Loko",
+  "!WassaW",
+  "!Nathan",
+  "!Zomzom",
+  "!Pellaeon",
+  "!Roystar",
+  "!Dikryl",
+  "!Alexis",
+  "!Ganapati",
+  "!Mad",
+  "!Mzboub",
+  "!Freez",
+];
+const cinvis = ["!membres", "!logs"];
+
+export { commandes, cpseudo, cinvis };
